ppp-patrol/requirements.txt
boto3
discord.py